+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.5.0 (TBD): Not yet released
 
 ### Minor features, changes, and enhancements
 [@GooberRF](https://github.com/GooberRF)
+- Notify players in a server that is recording demos
 - Restore cut first person weapon aim sway, toggleable with `cl_weaponsway`
 
 ### Bug fixes